Tests added.

+++ Semester_3/2.1/IViewer.cs
@@ -7,6 +7,6 @@ namespace Net
 	/// Interface for viewer.
 	public interface IViewer
 	{
-		void View(Net net, int iteration);
+		void View(Net net, ulong iteration);
 	}
 }
